pyproject.toml
[project]
name = "calendar-mcp"
version = "0.1.0"
description = "Google Calendar MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "google-api-python-client>=2.175.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "mcp>=1.10.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
